pyproject.toml
[tool.poetry]
name = "Langchain capstone project"
version = "0.1.0"
description = "Capstone project for Wizeline Langchain sprint"
authors = ["Roberto Galindo"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
numpy = "^1.24.3"
pandas = "^2.0.2"
scikit-learn = "^1.2.2"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
fastapi = "0.99.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
langchain = "^0.0.278"
tiktoken = "^0.4.0"
faiss-cpu = "^1.7.4"
chromadb = "^0.4.8"
ragas = "^0.0.11"
pypdf2 = "^3.0.1"
pypdf = "^3.15.4"
pdfminer-six = "^20221105"
python-multipart = "^0.0.6"



[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501", "B008"]
select = ["B","C","E","F","W","B9", "I", "N", "Q"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.mccabe]
max-complexity = 10
